+++ src/main/java/com/fj/SpringBootPlayground/service/MessageService.java
@@ -6,13 +6,13 @@ import com.fj.SpringBootPlayground.model.MessageModel;
 
 public interface MessageService {
 
-	public abstract MessageModel retrieveSingleMessage(int id);
+	public abstract MessageModel retrieveSingleMessage(Long id);
 
 	public abstract MessageModel addSingleMessage(MessageModel messageModel, boolean mapForInsert);
 
-	public abstract MessageModel updateMessage(int id, MessageModel messageModel);
+	public abstract MessageModel updateMessage(Long id, MessageModel messageModel);
 
-	public abstract MessageModel deleteSingleMessage(int id);
+	public abstract MessageModel deleteSingleMessage(Long id);
 
 	public abstract List<MessageModel> retrieveAllMessages();
